+++ CHANGELOG.md
@@ -11,8 +11,10 @@
   ellipsis returned by function `truncate-string-ellipsis`). To display the full content of a
   truncated cell in a dedicated buffer, type `v`.
   
-- The text button `Display procedures` in the main table-list buffer displays all the user-defined
-  functions and procedures in the current database.
+- The text button `Display procedures` in the main table-list buffer displays a table containing all
+  the user-defined functions and procedures in the current database.
+
+- The header line now shows whether the current database connection is secured with TLS.
 
 
 ## [0.12] - 2024-08-06